misc. optimizations

+++ headers/exercises.h
@@ -2,7 +2,7 @@
 #define  EXERCISES_H_
 #include "common.h"
 
-void _curls(float [][BUFFER_SIZE], int [][BUFFER_SIZE], unsigned int, bool buttonState, int);
+void _curls(int [][BUFFER_SIZE], int [][BUFFER_SIZE], unsigned int, bool buttonState, int);
 void _benchpress(int [][BUFFER_SIZE], unsigned int, int);
 void _squats(int [][BUFFER_SIZE], unsigned int, int);
 void _fitnessTest();
